requirements.txt
git+https://github.com/Rapptz/discord.py@rewrite
dnspython==1.16.0
PyNaCl==1.3.0
async-timeout==3.0.1
discord.py
random
string
asyncio
captcha.py